调整配置确保打包能正常启动
