+++ src/client/resources/nekowei-tool.client.mixins.json
@@ -3,7 +3,6 @@
 	"package": "top.nekowei.mc.mixin.client",
 	"compatibilityLevel": "JAVA_21",
 	"client": [
-		"ExampleClientMixin"
 	],
 	"injectors": {
 		"defaultRequire": 1
