+++ src/main/java/top/nekowei/mc/mixin/ExampleMixin.java
@@ -1,15 +1,15 @@
-package top.nekowei.mc.mixin;
-
-import net.minecraft.server.MinecraftServer;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-
-@Mixin(MinecraftServer.class)
-public class ExampleMixin {
-	@Inject(at = @At("HEAD"), method = "loadWorld")
-	private void init(CallbackInfo info) {
-		// This code is injected into the start of MinecraftServer.loadWorld()V
-	}
-}
+//package top.nekowei.mc.mixin;
+//
+//import net.minecraft.server.MinecraftServer;
+//import org.spongepowered.asm.mixin.Mixin;
+//import org.spongepowered.asm.mixin.injection.At;
+//import org.spongepowered.asm.mixin.injection.Inject;
+//import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+//
+//@Mixin(MinecraftServer.class)
+//public class ExampleMixin {
+//	@Inject(at = @At("HEAD"), method = "loadWorld")
+//	private void init(CallbackInfo info) {
+//		// This code is injected into the start of MinecraftServer.loadWorld()V
+//	}
+//}
