+++ src/main/resources/nekowei-tool.mixins.json
@@ -3,7 +3,6 @@
 	"package": "top.nekowei.mc.mixin",
 	"compatibilityLevel": "JAVA_21",
 	"mixins": [
-		"ExampleMixin"
 	],
 	"injectors": {
 		"defaultRequire": 1
